refactor(backend): feature-first validation, drop schemas/ folder
Kills the confusing db/schema.ts (Drizzle) vs schemas/ (Zod) naming collision.
Contest validation moves to r2/validation.ts (r2 = contest read, admin = contest
write, same domain); each feature now owns its own validation.ts going forward.
No behavior change — typecheck + all tests pass.
