+++ backend/src/r2/validation.ts
