--- backend/src/schemas/index.ts
@@ -1,3 +1,0 @@
-// Barrel for request/validation schemas. Add new domains as their own file
-// (e.g. ./forum, ./common) and re-export them here.
-export * from './r2';
